Chore: Small package name refactor
