+++ main.go
@@ -2,18 +2,12 @@ package main
 
 import (
 	"net"
-	"net/http"
 	"os"
 	"sync"
 
-	ChunkCPPGoAdapter "github.com/Sense-Scape/Go_TCP_Websocket_Adapter/v2/ChunkCPPGoAdapter"
+	"github.com/Sense-Scape/Go_TCP_Websocket_Adapter/v2/Routines"
+
 	"github.com/rs/zerolog"
-
-	"time"
-
-	"github.com/gin-contrib/cors"
-	"github.com/gin-gonic/gin"
-	"github.com/gorilla/websocket"
 )
 
 func main() {
@@ -23,7 +17,7 @@ func main() {
 	logger := zerolog.New(os.Stderr).With().Timestamp().Logger()
 	var wg sync.WaitGroup
 	wg.Add(1)
-	go handleConnectionWebSocketOutgoing(dataChannel, &wg)
+	go Routines.HandleConnectionGoWebSocketOutgoing(dataChannel, &wg)
 
 	// Define the port to listen on
 	port := "10005"
@@ -45,56 +39,8 @@ func main() {
 			logger.Error().Msg("Error:" + err.Error())
 			continue
 		}
-		go ChunkCPPGoAdapter.HandleConnectionTCPIncomingChunkTypes(dataChannel, conn)
+		go Routines.HandleConnectionTCPIncomingChunkTypes(dataChannel, conn)
 	}
 
 	wg.Wait()
-}
-
-var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
-}
-
-func handleConnectionWebSocketOutgoing(dataChannel <-chan string, wg *sync.WaitGroup) {
-	defer wg.Done()
-
-	logger := zerolog.New(os.Stderr).With().Timestamp().Logger()
-	logger.Info().Msg("TransportLayerDataSize:")
-	println("wohoo-----")
-
-	router := gin.Default()
-
-	// CORS middleware setup
-	corsConfig := cors.DefaultConfig()
-	corsConfig.AllowOrigins = []string{"http://localhost:5173/"} // Replace with your SvelteKit frontend URL
-
-	corsGroup := router.Group("/private")
-	corsGroup.Use(cors.New(corsConfig))
-
-	// Allow all origins to connect
-	// Note that is is not safe
-	upgrader.CheckOrigin = func(r *http.Request) bool {
-		return true
-	}
-
-	router.GET("/public", func(c *gin.Context) {
-
-		conn, err := upgrader.Upgrade(c.Writer, c.Request, nil)
-		if err != nil {
-			logger.Error().Msg(err.Error())
-			return
-		}
-		defer conn.Close()
-		for {
-			println("wohoo-----<")
-			dataString := <-dataChannel
-			println("wohoo-----<<")
-			conn.WriteMessage(websocket.TextMessage, []byte(dataString))
-			time.Sleep(time.Second)
-			logger.Debug().Msg("Transmitting Websocket data")
-		}
-	})
-	router.Run(":10010")
-
 }
